requirements.txt.txt
Flask==2.3.3
pymongo==4.5.0
bcrypt==4.0.1
PyJWT==2.8.0
pandas==2.1.1
openpyxl==3.1.2
groq==0.3.1
python-dotenv==1.0.0
Werkzeug==2.3.7